Addition of MBart EncoderLayer, definition of Pegasus EncoderLayer as MBartEncoderLayer
