+++ src/mbart/mod.rs
@@ -1,7 +1,10 @@
 mod attention;
 mod embeddings;
+mod encoder;
 mod mbart_model;
 
 pub use mbart_model::{
     MBartConfig, MBartConfigResources, MBartModelResources, MBartVocabResources,
 };
+
+pub(crate) use encoder::MBartEncoderLayer;
